# proyect2/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Wild Fantasy API",
    "description": "API documentation for managing characters, factions, places, species, items, and more in the Wild Fantasy universe.",
    "version": "1.0.0",
    "contact": {
      "name": "Joshua Burkhan"
    }
  },
  "host": "localhost:3000",
  "servers": [
    "http://localhost:3000"
  ],
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/api/newworld/abilities/": {
      "get": {
        "tags": [
          "Abilities"
        ],
        "description": "Retrieve all abilities",
        "responses": {
          "200": {
            "description": "Successfully retrieved abilities",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "example": "string"
                  },
                  "name": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "string"
                  },
                  "element": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "tags": [
          "Abilities"
        ],
        "description": "Create a new ability.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Ability data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "string"
                },
                "elements": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "element": {
                        "type": "string",
                        "example": "string"
                      },
                      "orbs": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                },
                "knownUsers": {
                  "type": "array",
                  "example": [
                    "string"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Ability created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "string"
                },
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "string"
                },
                "elements": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "element": {
                        "type": "string",
                        "example": "string"
                      },
                      "orbs": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                },
                "knownUsers": {
                  "type": "array",
                  "example": [
                    "string"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Error creating ability",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Error creating ability"
                },
                "error": {
                  "type": "object",
                  "properties": {}
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/newworld/abilities/{id}": {
      "get": {
        "tags": [
          "Abilities"
        ],
        "description": "Retrieve a specific ability by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Ability ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved ability"
          },
          "404": {
            "description": "Ability not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "tags": [
          "Abilities"
        ],
        "description": "Update an existing ability by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Ability ID"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated ability data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "string"
                },
                "elements": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "element": {
                        "type": "string",
                        "example": "string"
                      },
                      "orbs": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                },
                "knownUsers": {
                  "type": "array",
                  "example": [
                    "string"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ability updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "string"
                },
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "string"
                },
                "elements": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "element": {
                        "type": "string",
                        "example": "string"
                      },
                      "orbs": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                },
                "knownUsers": {
                  "type": "array",
                  "example": [
                    "string"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Error updating ability",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Error updating ability"
                },
                "error": {
                  "type": "object",
                  "properties": {}
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Ability not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Ability not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Abilities"
        ],
        "description": "Delete an ability by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Ability ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Ability deleted successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Ability deleted successfully"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Ability not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Ability not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error deleting ability",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Error deleting ability"
                },
                "error": {
                  "type": "object",
                  "properties": {}
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}